requirements.txt

Flask==3.0.0
pymongo[srv]
python-dotenv
gunicorn
